minor tweaks to api

+++ api-server/db/migrate/20170828153734_create_entries.rb
@@ -2,9 +2,9 @@ class CreateEntries < ActiveRecord::Migration[5.1]
   def change
     create_table :entries do |t|
       t.date :date
-      t.integer :distance_in_metre
+      t.float :distance_in_metre
       t.integer :time_in_second
-      t.integer :speed
+      t.float :speed
       t.references :user, index: true, foreign_key: true
 
       t.timestamps
